pyproject.toml
[project]
name = "flood-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pyinstaller>=6.15.0",
    "pyproj>=3.7.1",
    "pyside6>=6.9.1",
    "rasterio>=1.4.3",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "segmentation-models-pytorch>=0.5.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
